requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
pandas==2.2.0
pytest==7.4.4
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
pillow==10.2.0
boto3==1.34.40 